fix possible conversion from 'uint64_t' to 'size_t' warning on 32-bit builds
warning message:
c:\projects\openrw\external\microprofile\microprofile\microprofile.cpp(768): warning C4244: 'argument': conversion from 'uint64_t' to 'size_t', possible loss of data [C:\projects\openrw\build\external\microprofile\microprofile.vcxproj]

+++ microprofile.cpp
@@ -760,7 +760,7 @@ namespace
 		};
 	};
 };
-inline MicroProfileLogEntry MicroProfileMakeLogPayload(const char* pData, uint64_t nSize)
+inline MicroProfileLogEntry MicroProfileMakeLogPayload(const char* pData, size_t nSize)
 {
 	MP_ASSERT(nSize < 8);
 	MicroProfilePayloadPack P;
